+++ src/StudioCore/MsbEditor/Gizmos.cs
@@ -2,6 +2,7 @@
 using StudioCore.Renderer.DebugPrimitives;
 using StudioCore.Editor;
 using StudioCore.Renderer.Scene;
+using StudioCore.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Drawing;
