+++ src/StudioCore/MsbEditor/SpecialMapConnections.cs
@@ -1,5 +1,6 @@
 ﻿using Andre.Formats;
 using StudioCore.ParamEditor;
+using StudioCore.Utilities;
 using System;
 using System.Collections.Generic;
 using System.Linq;
