+++ src/StudioCore/Renderer/DebugPrimitives/DbgPrimSolidBone.cs
@@ -1,4 +1,5 @@
-﻿using System;
+﻿using StudioCore.Utilities;
+using System;
 using System.Drawing;
 using System.Numerics;
