+++ src/StudioCore/Renderer/DebugPrimitives/DbgPrimWireCylinder.cs
@@ -1,4 +1,5 @@
 ﻿using StudioCore.Renderer.Scene;
+using StudioCore.Utilities;
 using System;
 using System.Drawing;
 using System.Numerics;
