+++ src/StudioCore/Renderer/DebugPrimitives/DbgPrimWireWallBox.cs
@@ -1,4 +1,5 @@
 ﻿using StudioCore.Renderer.Scene;
+using StudioCore.Utilities;
 using System.Collections.Generic;
 using System.Drawing;
 using System.Numerics;
